♻️ Extract redux state update codes to hooks
