+++ hooks/index.ts
@@ -9,12 +9,14 @@ export { useFee } from './useFee'
 export { useMediaQuery } from './useMediaQuery'
 export { useModal } from './useModal'
 export { usePolling } from './usePolling'
+export { usePoolBalances } from './usePoolBalances'
 export { useProvider } from './useProvider'
 export { useAppDispatch, useAppSelector } from './useRedux'
 export { useReward } from './useReward'
 export { useStake } from './useStake'
 export { useTimer } from './useTimer'
 export { useToast } from './useToast'
+export { useTokenPrices } from './useTokenPrices'
 export { useTransaction } from './useTransaction'
 export { useUnstake } from './useUnstake'
 export { useUsd } from './useUsd'
